git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@12870 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/audio.h
@@ -387,7 +387,7 @@ class CAudio {
 			}
 		inline int MaxChannels (void) { return m_info.nMaxChannels; }
 		inline int FreeChannel (void) { return m_info.nFreeChannel; }
-		inline CAudioChannel* Channel (uint i = 0) { return (i < MaxChannels ()) ? m_channels + i : NULL; }
+		inline CAudioChannel* Channel (uint i = 0) { return (i < (uint) MaxChannels ()) ? m_channels + i : NULL; }
 		inline CStack<CSoundObject>& Objects (void) { return m_objects; }
 		inline float SlowDown (void) { return m_info.fSlowDown; }
 		inline void ActivateObject (void) { m_info.nActiveObjects++; }
